requirements.txt
Flask
Flask-SQLAlchemy
RPi.GPIO
dht11
apds9930
openai
